move main tests back and leave to separate CL
