--- src/cmd/go/internal/modfetch/main_test.go
@@ -1,29 +1,0 @@
-// Copyright 2018 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package modfetch
-
-import (
-	"io/ioutil"
-	"log"
-	"os"
-	"testing"
-
-	"cmd/go/internal/modfetch/codehost"
-)
-
-func TestMain(m *testing.M) {
-	os.Exit(testMain(m))
-}
-
-func testMain(m *testing.M) int {
-	dir, err := ioutil.TempDir("", "gitrepo-test-")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	codehost.WorkRoot = dir
-	return m.Run()
-}
