+++ src/cmd/go/internal/modload/query_test.go
@@ -6,12 +6,33 @@ package modload
 
 import (
 	"internal/testenv"
+	"io/ioutil"
+	"log"
+	"os"
 	"path"
+	"path/filepath"
 	"strings"
 	"testing"
 
+	"cmd/go/internal/modfetch"
+	"cmd/go/internal/modfetch/codehost"
 	"cmd/go/internal/module"
 )
+
+func TestMain(m *testing.M) {
+	os.Exit(testMain(m))
+}
+
+func testMain(m *testing.M) int {
+	dir, err := ioutil.TempDir("", "modload-test-")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer os.RemoveAll(dir)
+	modfetch.SrcMod = filepath.Join(dir, "src/mod")
+	codehost.WorkRoot = filepath.Join(dir, "codework")
+	return m.Run()
+}
 
 var (
 	queryRepo   = "vcs-test.golang.org/git/querytest.git"
